Removed Ansible prereq from doc/vagrant.md
Removed the unnecessary Ansible requirement from the Vagrant
documentation. This line was redundant because Ansible is prerequisite
for Ursula. This line was also confusing because it had listed Ansible
(1.7ish), which was incorrect since the current version of Ursula
requires 2.1.X.

Corrected the spelling of "prerequisites" in the same file.

Signed-off-by: Matt Langbehn <matthew.langbehn@gmail.com>